+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [2.4.0] - 2018-02-18
+### Changed
+- prototype.spawn.js, implements automatic settings for creeps tier 1
+- settings.js, calc automatic settings for creeps tier 1
+
 ## [2.7.5] - 2018-02-18
 ### Changed
 - prototype.spawn.js, fallback to Long Range Harvester